ft_strmapi.c

+++ CommonCore/Libft/ft_split.c
@@ -6,7 +6,7 @@
 /*   By: liferrei <liferrei@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/22 14:48:42 by liferrei          #+#    #+#             */
-/*   Updated: 2025/07/25 14:33:16 by liferrei         ###   ########.fr       */
+/*   Updated: 2025/07/25 14:40:59 by liferrei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,9 +48,9 @@ static void	ft_free_words(char **words, size_t n)
 
 static char	**ft_array_words(char const *s, int words, char c, char **array)
 {
-	size_t	i;
-	size_t	j;
-	size_t	start;
+	int	i;
+	int	j;
+	int	start;
 
 	i = 0;
 	j = 0;
@@ -90,6 +90,7 @@ char	**ft_split(char const *s, char c)
 	array_result = ft_array_words(s, words, c, array_result);
 	return (array_result);
 }
+
 /*
 #include <stdio.h>
 int main(void)
